tradeMapBack/tradeMapClone-main/api-flask-consulta-yahoo-b3/requirements.txt
astroid==2.4.2
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
lxml==4.5.0
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.4
pandas==1.1.4
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
requests-futures==1.0.0
selenium==3.141.0
six==1.15.0
toml==0.10.2
tqdm==4.52.0
typed-ast==1.4.1
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
yahooquery==2.2.8
kafka-python==2.2.1


